requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==24.3
python-dotenv==1.1.1
psycopg2-binary==2.9.10
dj-database-url==3.0.1
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
